Обновление и регистрация пользователя готова

+++ src/user/models/user.model.ts
@@ -56,7 +56,7 @@ export class User extends Model<User, UserModelAttrs> {
   @Column({ type: DataType.STRING, allowNull: true })
   banReason: string;
 
-  @Column({ type: DataType.INTEGER, allowNull: true })
+  @Column({ type: DataType.BIGINT, allowNull: true })
   lastUpdateEmail: number;
 
   // Уведомлять ли админа тг канала перед публикацией
